Re #3058 Muon Differences Clean up
Alphabetised test file names and merged workspacenaming test files into one
